pyproject.toml
[project]
name = "tavern"
version = "0.1.0"
description = "SRD 5e-compatible RPG engine with an LLM-powered narrator"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.88.0",
    "motor[srv]>=3.0.0",
    "discord.py>=2.4.0",
    "websockets>=13.0",
]

[dependency-groups]
discord-bot = [
    "discord.py>=2.3",
    "httpx>=0.27",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["backend/tavern/core/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tavern/tests"]
